+++ xen/include/asm-x86/hvm/vpt.h
@@ -188,7 +188,7 @@ void rtc_update_clock(struct domain *d);
 void pmtimer_init(struct vcpu *v);
 void pmtimer_deinit(struct domain *d);
 void pmtimer_reset(struct domain *d);
-int pmtimer_change_ioport(struct domain *d, unsigned int version);
+int pmtimer_change_ioport(struct domain *d, uint64_t version);
 
 void hpet_init(struct domain *d);
 void hpet_deinit(struct domain *d);
